Remove broken visit_each
The test for it is not run by the test suit, it does not compile because of
missing includes somewhere, and fixing compilation is not the end of the story.
The `visit_each` seems to be simply never worked, running the test fails with a
wrong result.

The only potential usage of `visit_each` is the Signals1/2 library, and I did
not find any bug report about that.